refactor: Update condition for checking if user has no address
The code in auth.reducer.ts has been refactored to update the condition for checking if a user has no address. The previous condition was not correctly accessing the addresses property of the user object. This commit fixes that issue by properly accessing the addresses property from the user data object.