add user address in location component (#52)
* first-commit

* style: format code

* chore: Update test-utils.ts

- Added a wait of 200ms before interacting with the address selector
- Changed the assertion for ion-popover visibility to check for existence instead

* chore: Remove unnecessary code in pap-report-ticket.cy.ts

This commit removes a block of code that verifies addresses from an API response. The code is no longer needed and has been removed to improve code cleanliness and maintainability.

* chore: Add delay before testing form behavior at second step

Added a 500ms delay before testing the correct behavior of the form at the second step in order to ensure proper synchronization.

* chore: Add delay before going to third step in pap-report-ticket.cy.ts

Add a 500ms delay before proceeding to the third step in the test for the correct behavior of the form at the second step in pap-report-ticket.cy.ts. This delay ensures that the necessary actions are completed before moving on to the next step.

* fix: fix arrow function syntax in pap-sign-up.cy.ts

The arrow function syntax was corrected in the test case for checking the label that matches one of the apiZonesGeoJson user types.

* chore: Update address selector component and location component

- Updated the class name in the address-selector.component.html template.
- Added a new CSS class for centering content in the address-selector.component.scss file.
- Added a new CSS class for styling the icon container in the address-selector.component.scss file.
- Dispatched an action to set marker coordinates to [0, 0] in the location.component.ts file.

* chore: Update address-selector.component.html

- Added *ngIf condition to display the "altro" option only if 'other' is true.

* test: pap settings and job test updated

* feat(select): remove unnecessary display of option ID

The commit removes the display of the option ID in the select component's HTML template. This change improves the user interface by only showing the option name.

* test: pap settings and test utils updated

* chore: Update info component template and logic

- Refactored the HTML template of the info component to simplify the structure.
- Updated the image source and page body content bindings in the template.
- Added an HTTP request to fetch the company page content from an API endpoint.
- Sanitized and displayed the fetched page content in the component.

This commit improves the readability and functionality of the info component.

* style: format code

* test: test refactor

---------

Co-authored-by: peppedeka <peppedeka@users.noreply.github.com>
Co-authored-by: Rubens Garofalo <rubensgiuseppegarofalo@gmail.com>