feat: Add missed user fields modal
This commit adds functionality to display a modal for missed user fields. When a user is logged in and there are required fields that have not been filled, the modal will be shown with the list of missed fields. After the user updates their profile, a success alert will be displayed.

- Added import statements for MissedFieldsUserModal and FormJson
- Added missedFields$ observable to select missed user fields from the store
- Implemented logic to show the MissedFieldsUserModal when required fields are missing
- Displayed a success alert after updating the profile
feat: Add missed user fields modal

This commit adds functionality to display a modal for missed user fields. When a user is logged in and there are required fields that have not been filled, the modal will be shown with the list of missed fields. After the user updates their profile, a success alert will be displayed.

- Added import statements for MissedFieldsUserModal and FormJson
- Added missedFields$ observable to select missed user fields from the store
- Implemented logic to show the MissedFieldsUserModal when required fields are missing
- Displayed a success alert after updating the profile