fix: Fix date formatting in push notification page component
The commit fixes an issue where the date was not being formatted correctly in the push notification page component. The fix ensures that the date is displayed in the format "dd/MM/yyyy hh:mm a".

chore: Update push-notification-page.component.html

- Updated the date format to display the updated_at value instead of schedule_date or created_at.

chore: Unsubscribe from pushNotificationsSub in ngOnDestroy

This commit adds an unsubscribe method for the _pushNotificationsSub subscription in the ngOnDestroy lifecycle hook of the PushNotificationsPageComponent. This ensures that the subscription is properly unsubscribed when the component is destroyed, preventing memory leaks.